refactor(reports): move shared report data model out of email_template into report_data
NoteworthyEvent / ProjectReportData / ReportData are the shared report data model
rendered by BOTH the email templates and the Slack report formatter, but they lived
in reports/email_template.rs — so Slack code imported report types from a file named
"email_template" (misleading; the file held no template, only the structs).

Rename email_template.rs -> report_data.rs, re-export the three types from reports,
and point all consumers at crate::reports::{...}. Pure move + import update, no
behavior change.
